requirements.txt
# ==============================================================
#      متطلبات مشروع المتجر الإلكتروني الجزائري (Amazon-Like)
#      تاريخ الإنشاء: 2026
#      البيئة: Python 3.12+
# ==============================================================


# ---------------------------------------------------------------
# 1. الإطار الأساسي - Core Framework
# ---------------------------------------------------------------

Django==5.1.4
django-jazzmin==3.0.1


# ---------------------------------------------------------------
# 2. واجهات برمجة التطبيقات - REST API
# ---------------------------------------------------------------

djangorestframework==3.15.2


# ---------------------------------------------------------------
# 3. معالجة الصور - Image Processing
# ---------------------------------------------------------------

Pillow==11.0.0


# ---------------------------------------------------------------
# 4. إدارة CORS - Cross-Origin Resource Sharing
# ---------------------------------------------------------------

django-cors-headers==4.6.0


# ---------------------------------------------------------------
# 5. إدارة المتغيرات البيئية - Environment Variables
# ---------------------------------------------------------------

python-dotenv==1.0.1


# ---------------------------------------------------------------
# 6. قاعدة البيانات - Database Adapters (اختر حسب حاجتك)
# ---------------------------------------------------------------

# SQLite مضمن مع Python — لا يحتاج تثبيت إضافي (الافتراضي)
# mysqlclient==2.2.7         # فعّل هذا عند النقل إلى MySQL
# psycopg2-binary==2.9.10   # فعّل هذا عند النقل إلى PostgreSQL


# ---------------------------------------------------------------
# 7. الفلترة المتقدمة - Advanced Filtering
# ---------------------------------------------------------------

django-filter==24.3


# ==============================================================
# مكتبات إضافية مُقترحة للمراحل القادمة (اختياري الآن)
# ==============================================================

# celery==5.4.0            # لتنفيذ المهام في الخلفية
# redis==5.2.1             # Broker لـ Celery وCache
# stripe==11.4.1           # بوابة الدفع الإلكتروني الدولية
# django-storages==1.14.4  # رفع الملفات على السحابة
# gunicorn==23.0.0         # WSGI Server للإنتاج